Renamed xrGameMtlLib to xrMaterialSystem
