+++ src/engine.sln
@@ -215,7 +215,7 @@ Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "xrLuaFix", "..\Externals\xr
 EndProject
 Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "DiscordGameSDK", "..\Externals\DiscordGameSDK\DiscordGameSDK.vcxproj", "{CFF9F0DD-C2FC-424C-800B-BFBA35003932}"
 EndProject
-Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "xrGameMtlLib", "xrGameMtlLib\xrGameMtlLib.vcxproj", "{2C419512-6EEE-4707-BC51-2E834855552E}"
+Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "xrMaterialSystem", "xrMaterialSystem\xrMaterialSystem.vcxproj", "{2C419512-6EEE-4707-BC51-2E834855552E}"
 	ProjectSection(ProjectDependencies) = postProject
 		{2578C6D8-660D-48AE-9322-7422F8664F06} = {2578C6D8-660D-48AE-9322-7422F8664F06}
 		{A0F7D1FB-59A7-4717-A7E4-96F37E91998E} = {A0F7D1FB-59A7-4717-A7E4-96F37E91998E}
@@ -2082,10 +2082,10 @@ Global
 		{CFF9F0DD-C2FC-424C-800B-BFBA35003932}.Release|x64.Build.0 = Release|x64
 		{CFF9F0DD-C2FC-424C-800B-BFBA35003932}.Release|x86.ActiveCfg = Release|Win32
 		{CFF9F0DD-C2FC-424C-800B-BFBA35003932}.Release|x86.Build.0 = Release|Win32
-		{2C419512-6EEE-4707-BC51-2E834855552E}.Debug|ARM.ActiveCfg = Debug|Win32
-		{2C419512-6EEE-4707-BC51-2E834855552E}.Debug|ARM.Build.0 = Debug|Win32
-		{2C419512-6EEE-4707-BC51-2E834855552E}.Debug|ARM64.ActiveCfg = Debug|x64
-		{2C419512-6EEE-4707-BC51-2E834855552E}.Debug|ARM64.Build.0 = Debug|x64
+		{2C419512-6EEE-4707-BC51-2E834855552E}.Debug|ARM.ActiveCfg = Debug|ARM
+		{2C419512-6EEE-4707-BC51-2E834855552E}.Debug|ARM.Build.0 = Debug|ARM
+		{2C419512-6EEE-4707-BC51-2E834855552E}.Debug|ARM64.ActiveCfg = Debug|ARM64
+		{2C419512-6EEE-4707-BC51-2E834855552E}.Debug|ARM64.Build.0 = Debug|ARM64
 		{2C419512-6EEE-4707-BC51-2E834855552E}.Debug|x64.ActiveCfg = Debug|x64
 		{2C419512-6EEE-4707-BC51-2E834855552E}.Debug|x64.Build.0 = Debug|x64
 		{2C419512-6EEE-4707-BC51-2E834855552E}.Debug|x86.ActiveCfg = Debug|Win32
@@ -2106,10 +2106,10 @@ Global
 		{2C419512-6EEE-4707-BC51-2E834855552E}.Release Master Gold|x64.Build.0 = Release Master Gold|x64
 		{2C419512-6EEE-4707-BC51-2E834855552E}.Release Master Gold|x86.ActiveCfg = Release Master Gold|Win32
 		{2C419512-6EEE-4707-BC51-2E834855552E}.Release Master Gold|x86.Build.0 = Release Master Gold|Win32
-		{2C419512-6EEE-4707-BC51-2E834855552E}.Release|ARM.ActiveCfg = Release|Win32
-		{2C419512-6EEE-4707-BC51-2E834855552E}.Release|ARM.Build.0 = Release|Win32
-		{2C419512-6EEE-4707-BC51-2E834855552E}.Release|ARM64.ActiveCfg = Release|x64
-		{2C419512-6EEE-4707-BC51-2E834855552E}.Release|ARM64.Build.0 = Release|x64
+		{2C419512-6EEE-4707-BC51-2E834855552E}.Release|ARM.ActiveCfg = Release|ARM
+		{2C419512-6EEE-4707-BC51-2E834855552E}.Release|ARM.Build.0 = Release|ARM
+		{2C419512-6EEE-4707-BC51-2E834855552E}.Release|ARM64.ActiveCfg = Release|ARM64
+		{2C419512-6EEE-4707-BC51-2E834855552E}.Release|ARM64.Build.0 = Release|ARM64
 		{2C419512-6EEE-4707-BC51-2E834855552E}.Release|x64.ActiveCfg = Release|x64
 		{2C419512-6EEE-4707-BC51-2E834855552E}.Release|x64.Build.0 = Release|x64
 		{2C419512-6EEE-4707-BC51-2E834855552E}.Release|x86.ActiveCfg = Release|Win32
@@ -2162,7 +2162,6 @@ Global
 		{C8FBC3CE-D6DE-4FC3-BC15-7B647614DB09} = {2BFC806B-CE92-4EA4-8FE8-5F2EA54BA348}
 		{4E076FEB-DC78-4BF7-99C7-AC8267247989} = {49438080-78B8-4056-BDCB-4DACAD652C21}
 		{CFF9F0DD-C2FC-424C-800B-BFBA35003932} = {2BFC806B-CE92-4EA4-8FE8-5F2EA54BA348}
-		{2C419512-6EEE-4707-BC51-2E834855552E} = {49438080-78B8-4056-BDCB-4DACAD652C21}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {C5A32302-94F5-41B8-9606-017BB6BA69A6}
