+++ src/xrGame/Actor_Feel.cpp
@@ -6,7 +6,7 @@
 #include "character_info.h"
 #include "xrEngine/xr_level_controller.h"
 #include "CustomZone.h"
-#include "xrGameMtlLib/GameMtlLib.h"
+#include "xrMaterialSystem/GameMtlLib.h"
 #include "ui/UIMainIngameWnd.h"
 #include "UIGameCustom.h"
 #include "Grenade.h"
