+++ src/xrGame/Explosive.cpp
@@ -19,7 +19,7 @@
 #include "Level.h"
 #include "Level_Bullet_Manager.h"
 #include "xrMessages.h"
-#include "xrGameMtlLib/GameMtlLib.h"
+#include "xrMaterialSystem/GameMtlLib.h"
 
 #ifdef DEBUG
 #include "xrEngine/StatGraph.h"
