+++ src/xrGame/HUDTarget.cpp
@@ -1,6 +1,6 @@
 #include "StdAfx.h"
 #include "HUDTarget.h"
-#include "xrGameMtlLib/GameMtlLib.h"
+#include "xrMaterialSystem/GameMtlLib.h"
 
 #include "xrEngine/Environment.h"
 #include "xrEngine/CustomHUD.h"
