+++ src/xrGame/wallmark_manager.cpp
@@ -1,7 +1,7 @@
 #include "StdAfx.h"
 #include "wallmark_manager.h"
 #include "Level.h"
-#include "xrGameMtlLib/GameMtlLib.h"
+#include "xrMaterialSystem/GameMtlLib.h"
 #include "xrPhysics/CalculateTriangle.h"
 #include "xrEngine/profiler.h"
 #ifdef DEBUG
