+++ src/xrGame/xrGame.vcxproj
@@ -3531,7 +3531,7 @@
     <ProjectReference Include="..\..\Externals\imgui.vcxproj">
       <Project>{c8fbc3ce-d6de-4fc3-bc15-7b647614db09}</Project>
     </ProjectReference>
-    <ProjectReference Include="..\xrGameMtlLib\xrGameMtlLib.vcxproj">
+    <ProjectReference Include="$(SolutionDir)xrMaterialSystem\xrMaterialSystem.vcxproj">
       <Project>{2c419512-6eee-4707-bc51-2e834855552e}</Project>
     </ProjectReference>
   </ItemGroup>
