+++ src/xrMaterialSystem/GameMtlLib.cpp
@@ -2,22 +2,6 @@
 #include "GameMtlLib.h"
 #include "Common/FSMacros.hpp"
 
-BOOL APIENTRY DllMain(HMODULE hModule,
-    DWORD  ul_reason_for_call,
-    LPVOID lpReserved
-)
-{
-    switch (ul_reason_for_call)
-    {
-    case DLL_PROCESS_ATTACH:
-    case DLL_THREAD_ATTACH:
-    case DLL_THREAD_DETACH:
-    case DLL_PROCESS_DETACH:
-        break;
-    }
-    return TRUE;
-}
-
 CGameMtlLibrary GMLib;
 #ifdef _EDITOR
 CGameMtlLibrary* PGMLib = nullptr;
