+++ src/xrMaterialSystem/GameMtlLib.h
