+++ src/xrMaterialSystem/GameMtlLib_Engine.cpp
