+++ src/xrMaterialSystem/stdafx.cpp
